three specs passing

+++ js/solarAgeCalculator-interface.js
@@ -3,7 +3,7 @@ import { Solar } from './../js/solarAgeCalculator.js';
 $(document).ready(function() {
   $("form#enterAge").submit(function(event) {
     event.preventDefault();
-    var age = parseInt($("#age").val());
+    var age = parseFloat($("#age").val());
 
     var ageChecker = new Solar(age);
